week9/requirements.txt
attrs==21.4.0
avro-python3==1.10.0
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
configparser==5.2.0
cryptography==37.0.2
fastavro==1.4.10
google-api-core==2.8.0
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
jsonschema==4.4.0
numpy==1.21.6
pandas==1.3.5
protobuf==3.20.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
